+++ src/controllers/highlights/HighlightBadge.cpp
@@ -8,26 +8,29 @@ QColor HighlightBadge::FALLBACK_HIGHLIGHT_COLOR = QColor(127, 63, 73, 127);
 
 bool HighlightBadge::operator==(const HighlightBadge &other) const
 {
-    return std::tie(this->badgeName_, this->displayName_, this->hasSound_,
-                    this->hasAlert_, this->soundUrl_, this->color_) ==
-           std::tie(other.badgeName_, other.displayName_, other.hasSound_,
-                    other.hasAlert_, other.soundUrl_, other.color_);
+    return std::tie(this->badgeName_, this->showInMentions_, this->displayName_,
+                    this->hasSound_, this->hasAlert_, this->soundUrl_,
+                    this->color_) ==
+           std::tie(other.badgeName_, other.showInMentions_, other.displayName_,
+                    other.hasSound_, other.hasAlert_, other.soundUrl_,
+                    other.color_);
 }
 
-HighlightBadge::HighlightBadge(const QString &badgeName,
+HighlightBadge::HighlightBadge(const QString &badgeName, bool showInMentions,
                                const QString &displayName, bool hasAlert,
                                bool hasSound, const QString &soundUrl,
                                QColor color)
-    : HighlightBadge(badgeName, displayName, hasAlert, hasSound, soundUrl,
-                     std::make_shared<QColor>(color))
+    : HighlightBadge(badgeName, showInMentions, displayName, hasAlert, hasSound,
+                     soundUrl, std::make_shared<QColor>(color))
 {
 }
 
-HighlightBadge::HighlightBadge(const QString &badgeName,
+HighlightBadge::HighlightBadge(const QString &badgeName, bool showInMentions,
                                const QString &displayName, bool hasAlert,
                                bool hasSound, const QString &soundUrl,
                                std::shared_ptr<QColor> color)
     : badgeName_(badgeName)
+    , showInMentions_(showInMentions)
     , displayName_(displayName)
     , hasAlert_(hasAlert)
     , hasSound_(hasSound)
@@ -46,6 +49,11 @@ HighlightBadge::HighlightBadge(const QString &badgeName,
 const QString &HighlightBadge::badgeName() const
 {
     return this->badgeName_;
+}
+
+bool HighlightBadge::showInMentions() const
+{
+    return this->showInMentions_;
 }
 
 const QString &HighlightBadge::displayName() const